Split profile CMS renderer and viewer modules
