+++ scripts/debt-ratchet-baseline.json
@@ -4,7 +4,7 @@
   "counts": {
     "any_casts": 4,
     "todo_comments": 0,
-    "oversized_files": 87,
+    "oversized_files": 85,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
@@ -76,8 +76,6 @@
     "components/nft-transfer/TransferModal.tsx",
     "components/notifications/NotificationsContext.tsx",
     "components/profile-cms-builder/ProfileCmsBuilder.tsx",
-    "components/profile-cms/CmsSiteRenderer.tsx",
-    "components/profile-cms/CmsThreeDViewer.tsx",
     "components/react-query-wrapper/ReactQueryWrapper.tsx",
     "components/rememes/RememePage.tsx",
     "components/user/collected/UserPageCollected.tsx",
